grpc-hooks-server/pyproject.toml
[project]
name = "grpc-hooks-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio-tools>=1.68.1",
    "pika>=1.3.2",
    "python-dotenv>=1.0.1",
]
